The rename, mappers & transformations separated
Trying to introduce some structure here, and separation between mappers
& transformations.
