+++ ruby_event_store/lib/ruby_event_store.rb
@@ -15,24 +15,27 @@ require 'ruby_event_store/stream'
 require 'ruby_event_store/expected_version'
 require 'ruby_event_store/serialized_record'
 require 'ruby_event_store/transform_keys'
-require 'ruby_event_store/mappers/transformation_item'
-require 'ruby_event_store/mappers/domain_event_mapper'
-require 'ruby_event_store/mappers/serialization_mapper'
-require 'ruby_event_store/mappers/serialized_record_mapper'
-require 'ruby_event_store/mappers/symbolize_metadata_keys'
-require 'ruby_event_store/mappers/stringify_metadata_keys'
-require 'ruby_event_store/mappers/event_class_remapper'
-require 'ruby_event_store/mappers/encryption'
-require 'ruby_event_store/mappers/pipeline_mapper'
+require 'ruby_event_store/mappers/encryption_key'
+require 'ruby_event_store/mappers/in_memory_encryption_key_repository'
+require 'ruby_event_store/mappers/transformation/domain_event'
+require 'ruby_event_store/mappers/transformation/encryption'
+require 'ruby_event_store/mappers/transformation/event_class_remapper'
+require 'ruby_event_store/mappers/transformation/item'
+require 'ruby_event_store/mappers/transformation/proto_event'
+require 'ruby_event_store/mappers/transformation/protobuf_encoder'
+require 'ruby_event_store/mappers/transformation/protobuf_nested_struct_metadata'
+require 'ruby_event_store/mappers/transformation/serialization'
+require 'ruby_event_store/mappers/transformation/serialized_record'
+require 'ruby_event_store/mappers/transformation/stringify_metadata_keys'
+require 'ruby_event_store/mappers/transformation/symbolize_metadata_keys'
 require 'ruby_event_store/mappers/pipeline'
+require 'ruby_event_store/mappers/pipeline_mapper'
 require 'ruby_event_store/mappers/default'
-require 'ruby_event_store/mappers/domain_event_proto_mapper'
-require 'ruby_event_store/mappers/proto_mapper'
-require 'ruby_event_store/mappers/protobuf_nested_struct_metadata_mapper'
-require 'ruby_event_store/mappers/protobuf'
-require 'ruby_event_store/mappers/null_mapper'
-require 'ruby_event_store/mappers/instrumented_mapper'
+require 'ruby_event_store/mappers/forgotten_data'
 require 'ruby_event_store/mappers/encryption_mapper'
+require 'ruby_event_store/mappers/instrumented_mapper'
+require 'ruby_event_store/mappers/null_mapper'
+require 'ruby_event_store/mappers/protobuf'
 require 'ruby_event_store/batch_enumerator'
 require 'ruby_event_store/correlated_commands'
 require 'ruby_event_store/link_by_metadata'
