--- ruby_event_store/lib/ruby_event_store/mappers/domain_event_proto_mapper.rb
@@ -1,13 +1,0 @@
-module RubyEventStore
-  module Mappers
-    class DomainEventProtoMapper < DomainEventMapper
-      def load(item)
-        Proto.new(
-          event_id: item.event_id,
-          data:     item.data,
-          metadata: item.metadata
-        )
-      end
-    end
-  end
-end
