+++ ruby_event_store/lib/ruby_event_store/mappers/encryption_mapper.rb
@@ -4,8 +4,8 @@ module RubyEventStore
       def initialize(key_repository, serializer: YAML, forgotten_data: ForgottenData.new)
         super(Pipeline.new(
           transformations: [
-            Encryption.new(key_repository, serializer: serializer, forgotten_data: forgotten_data),
-            SerializationMapper.new(serializer: serializer),
+            Transformation::Encryption.new(key_repository, serializer: serializer, forgotten_data: forgotten_data),
+            Transformation::Serialization.new(serializer: serializer),
           ]
         ))
       end
