+++ ruby_event_store/lib/ruby_event_store/mappers/pipeline.rb
@@ -1,8 +1,8 @@
 module RubyEventStore
   module Mappers
     class Pipeline
-      def initialize(to_domain_event: DomainEventMapper.new,
-                     to_serialized_record: SerializedRecordMapper.new,
+      def initialize(to_domain_event: Transformation::DomainEvent.new,
+                     to_serialized_record: Transformation::SerializedRecord.new,
                      transformations: nil)
         @transformations = [
           to_domain_event,
