--- ruby_event_store/lib/ruby_event_store/mappers/stringify_metadata_keys.rb
@@ -1,20 +1,0 @@
-module RubyEventStore
-  module Mappers
-    class StringifyMetadataKeys
-      def dump(item)
-        stringify(item)
-      end
-
-      def load(item)
-        stringify(item)
-      end
-
-      private
-      def stringify(item)
-        item.merge(
-          metadata: TransformKeys.stringify(item.metadata),
-        )
-      end
-    end
-  end
-end
