--- ruby_event_store/lib/ruby_event_store/mappers/symbolize_metadata_keys.rb
@@ -1,20 +1,0 @@
-module RubyEventStore
-  module Mappers
-    class SymbolizeMetadataKeys
-      def dump(item)
-        symbolize(item)
-      end
-
-      def load(item)
-        symbolize(item)
-      end
-
-      private
-      def symbolize(item)
-        item.merge(
-          metadata: TransformKeys.symbolize(item.metadata),
-        )
-      end
-    end
-  end
-end
