+++ ruby_event_store/lib/ruby_event_store/mappers/transformation/proto_event.rb
@@ -1,0 +1,15 @@
+module RubyEventStore
+  module Mappers
+    module Transformation
+      class ProtoEvent < DomainEvent
+        def load(item)
+          Proto.new(
+            event_id: item.event_id,
+            data:     item.data,
+            metadata: item.metadata
+          )
+        end
+      end
+    end
+  end
+end
