+++ ruby_event_store/lib/ruby_event_store/mappers/transformation/symbolize_metadata_keys.rb
@@ -1,0 +1,22 @@
+module RubyEventStore
+  module Mappers
+    module Transformation
+      class SymbolizeMetadataKeys
+        def dump(item)
+          symbolize(item)
+        end
+
+        def load(item)
+          symbolize(item)
+        end
+
+        private
+        def symbolize(item)
+          item.merge(
+            metadata: TransformKeys.symbolize(item.metadata),
+          )
+        end
+      end
+    end
+  end
+end
