+++ ruby_event_store/spec/mappers/pipeline_spec.rb
@@ -6,7 +6,7 @@ module RubyEventStore
 
       specify '#initialize - default values' do
         pipe = Pipeline.new
-        expect(pipe.transformations.map(&:class)).to eq [DomainEventMapper, SerializedRecordMapper]
+        expect(pipe.transformations.map(&:class)).to eq [Transformation::DomainEvent, Transformation::SerializedRecord]
       end
 
       specify '#initialize - custom edge mappers' do
@@ -39,15 +39,15 @@ module RubyEventStore
       end
 
       specify '#dump' do
-        domain_mapper = DomainEventMapper.new
-        record_mapper = SerializedRecordMapper.new
-        some_mapper1 = SymbolizeMetadataKeys.new
-        some_mapper2 = StringifyMetadataKeys.new
+        domain_mapper = Transformation::DomainEvent.new
+        record_mapper = Transformation::SerializedRecord.new
+        some_mapper1 = Transformation::SymbolizeMetadataKeys.new
+        some_mapper2 = Transformation::StringifyMetadataKeys.new
         pipe = Pipeline.new(to_domain_event: domain_mapper, to_serialized_record: record_mapper, transformations: [some_mapper1, some_mapper2])
         domain_event = TestEvent.new
-        item1 = TransformationItem.new(event_id: domain_event.event_id, item: 1)
-        item2 = TransformationItem.new(event_id: domain_event.event_id, item: 2)
-        item3 = TransformationItem.new(event_id: domain_event.event_id, item: 3)
+        item1 = Transformation::Item.new(event_id: domain_event.event_id, item: 1)
+        item2 = Transformation::Item.new(event_id: domain_event.event_id, item: 2)
+        item3 = Transformation::Item.new(event_id: domain_event.event_id, item: 3)
         expect(domain_mapper).to receive(:dump).with(domain_event).and_return(item1)
         expect(some_mapper1).to receive(:dump).with(item1).and_return(item2)
         expect(some_mapper2).to receive(:dump).with(item2).and_return(item3)
@@ -56,15 +56,15 @@ module RubyEventStore
       end
 
       specify '#dump' do
-        domain_mapper = DomainEventMapper.new
-        record_mapper = SerializedRecordMapper.new
-        some_mapper1 = SymbolizeMetadataKeys.new
-        some_mapper2 = StringifyMetadataKeys.new
+        domain_mapper = Transformation::DomainEvent.new
+        record_mapper = Transformation::SerializedRecord.new
+        some_mapper1 = Transformation::SymbolizeMetadataKeys.new
+        some_mapper2 = Transformation::StringifyMetadataKeys.new
         pipe = Pipeline.new(to_domain_event: domain_mapper, to_serialized_record: record_mapper, transformations: [some_mapper1, some_mapper2])
         record  = SerializedRecord.new(event_id: SecureRandom.uuid, data: '', metadata: '', event_type: 'TestEvent')
-        item1 = TransformationItem.new(event_id: record.event_id, item: 1)
-        item2 = TransformationItem.new(event_id: record.event_id, item: 2)
-        item3 = TransformationItem.new(event_id: record.event_id, item: 3)
+        item1 = Transformation::Item.new(event_id: record.event_id, item: 1)
+        item2 = Transformation::Item.new(event_id: record.event_id, item: 2)
+        item3 = Transformation::Item.new(event_id: record.event_id, item: 3)
         expect(record_mapper).to receive(:load).with(record).and_return(item1)
         expect(some_mapper2).to receive(:load).with(item1).and_return(item2)
         expect(some_mapper1).to receive(:load).with(item2).and_return(item3)
