+++ ruby_event_store/spec/mappers/transformation/symbolize_metadata_keys_spec.rb
@@ -1,0 +1,40 @@
+require 'spec_helper'
+require 'json'
+
+module RubyEventStore
+  module Mappers
+    module Transformation
+      RSpec.describe SymbolizeMetadataKeys do
+        let(:uuid)  { SecureRandom.uuid }
+        let(:item)  {
+          Item.new(
+            event_id:   uuid,
+            metadata:   {some: 'meta'},
+            data:       JSON.parse(JSON.dump({some: 'value'})),
+            event_type: 'TestEvent',
+          )
+        }
+        let(:changed_item)  {
+          Item.new(
+            event_id:   uuid,
+            metadata:   JSON.parse(JSON.dump({some: 'meta'})),
+            data:       JSON.parse(JSON.dump({some: 'value'})),
+            event_type: 'TestEvent',
+          )
+        }
+
+        specify "#dump" do
+          result = SymbolizeMetadataKeys.new.dump(changed_item)
+          expect(result).to eq(item)
+          expect(result[:metadata].keys.map(&:class).uniq).to eq([Symbol])
+        end
+
+        specify "#load" do
+          result = SymbolizeMetadataKeys.new.load(changed_item)
+          expect(result).to eq(item)
+          expect(result[:metadata].keys.map(&:class).uniq).to eq([Symbol])
+        end
+      end
+    end
+  end
+end
